Импорт данных из приложения "Журнал Вконтакте"

+++ database/migrations/2016_01_17_140850_create_questionnaires_table.php
@@ -17,7 +17,7 @@ class CreateQuestionnairesTable extends Migration {
 			$table->increments('id');
 			$table->char('student_id', 8);
 			$table->unsignedInteger('teacher_id');
-			$table->tinyInteger('academic_year')->unsigned();
+			$table->mediumInteger('academic_year')->unsigned();
 			$table->tinyInteger('semester')->unsigned();
 			$table->foreign('student_id')->references('id')->on('students');
 			$table->foreign('teacher_id')->references('id')->on('teachers');
